Moved Cache to separate folder
