+++ engine/common/follower/cache/cache.go
@@ -1,4 +1,4 @@
-package follower
+package cache
 
 import (
 	"sync"
